requirements.txt
discord.py==2.4.0
private==0.2.0